+++ CHANGELOG.rst
@@ -10,6 +10,8 @@
   Add toBase64 and fromBase64 to SeriDeseri
 * `#376 <https://github.com/xitrum-framework/xitrum/issues/376>`_
   Embed Xitrum CSRF token to requests from Swagger UI
+* `#379 <https://github.com/xitrum-framework/xitrum/issues/379>`_
+  Add xitrum.version
 
 3.11: